Created Order component
