+++ src/Restaurant.js
@@ -1,25 +1,14 @@
-import React, { useState } from "react";
-import RestaurantButton from "./RestaurantButton";
+import React from "react";
+
 import "./Restaurant.css";
+import Order from "./Order";
+
 const Restaurant = () => {
-  // array destructuring
-  // setOrders is a function that allows me to modify the value of the orders variable
-  // 0 is the initial value
-  const [orderCount, setOrderCount] = useState(0);
-
-  function orderOne() {
-    setOrderCount((orderCount) => {
-      return orderCount + 1;
-    });
-  }
-
   return (
     <div className="Restaurant">
       <h3>Restaurant Orders</h3>
       <ul>
-        <li>
-          Pizzas: {orderCount} <RestaurantButton orderOne={orderOne} />
-        </li>
+        <Order />
       </ul>
     </div>
   );
